Initial commit.
Provides two commands, switch and check, and very little else.

 -- switch <A NAME>

Records an entry with the current time as a timestamp

-- check <Number of Hours>

Checks if any of the names do not have a timestamp within the last X
hours and notifies the email addresses in config using the template
defined there.

Doesn't yet:

* Handle the 'failed on first run' case.

* Have pretty templates.

* Know where to send the email.